Issue 57

+++ backend/app/controllers/posts_controller.rb
@@ -18,25 +18,39 @@ class PostsController < ApplicationController
     end
 
     def create
-        post = Post.new(post_params)
-        post.image.attach(params[:post][:image]) if params[:post][:image].present?
-        if post.save
-            render json: post_attributes(post), status: :created
+        # メールアドレスがtest@example.comのユーザーの投稿を制限
+        if current_user.email == 'test@example.com'
+          render json: { error: 'testユーザーには投稿権限がありません' }, status: :forbidden
         else
-            render json: post.errors, status: :unprocessable_entity
+            post = Post.new(post_params)
+            post.image.attach(params[:post][:image]) if params[:post][:image].present?
+      
+            if post.save
+                render json: post_attributes(post), status: :created
+            else
+                render json: post.errors, status: :unprocessable_entity
+            end
+        end
+      end
+
+    def update
+        if current_user.email == 'test@example.com'
+            render json: { error: 'testユーザーには編集権限がありません' }, status: :forbidden
+        else
+            post = Post.find params[:id]
+            post.update(post_params)
+            render json: post_attributes(post)
         end
     end
 
-    def update
-        post = Post.find params[:id]
-        post.update(post_params)
-        render json: post_attributes(post)
-    end
-
     def destroy
-        post = Post.find params[:id]
-        post.destroy
-        render json: post
+        if current_user.email == 'test@example.com'
+            render json: { error: 'testユーザーには削除権限がありません' }, status: :forbidden
+        else
+            post = Post.find params[:id]
+            post.destroy
+            render json: post
+        end
     end
 
     def search
